Chore: Reorganize directory structure
