+++ packages/extension-browser/src/content-script/formatter.ts
@@ -1,7 +1,7 @@
 import { Category } from 'hint/dist/src/lib/enums/category';
 import { FormatterOptions, HintResources, IFormatter, IHintConstructor, Problem, Severity } from 'hint/dist/src/lib/types';
 
-import browser from './util/browser';
+import browser from '../shared/browser';
 
 export default class WebExtensionFormatter implements IFormatter {
