+++ packages/extension-browser/src/content-script/webhint.ts
