--- packages/extension-browser/src/devtools.ts
@@ -1,3 +1,0 @@
-import browser from './util/browser';
-
-browser.devtools.panels.create('Hints', 'icon.png', 'webhint-panel.html');
