+++ packages/extension-browser/src/devtools/devtools.html
