+++ packages/extension-browser/src/devtools/devtools.ts
@@ -1,0 +1,3 @@
+import browser from '../shared/browser';
+
+browser.devtools.panels.create('Hints', 'icon.png', '/devtools/panel/panel.html');
