+++ packages/extension-browser/src/devtools/panel/panel.html
@@ -12,6 +12,6 @@
         <h1>webhint</h1>
         <button id="toggle">Start</button>
         <pre id="results"></pre>
-        <script src="webhint-panel.js"></script>
+        <script src="panel.js"></script>
     </body>
 </html>
