+++ packages/extension-browser/src/manifest.json
@@ -14,7 +14,7 @@
         "180": "icon.png"
     },
     "description": "Run webhint in your browser.",
-    "devtools_page": "devtools-page.html",
+    "devtools_page": "devtools/devtools.html",
     "manifest_version": 2,
     "name": "webhint",
     "permissions": [
