+++ packages/extension-browser/src/shared/browser.ts
