+++ packages/extension-browser/src/shared/types.ts
